+++ src/amount.ts
