+++ src/index.ts
@@ -1,10 +1,9 @@
-export {
-  Amount,
-  Unit,
-  UnitFormat,
-  Format,
-  Serialize,
-  UnitPrefix,
-  BaseUnits
-} from "./core";
-// export { Units, UnitsFormat, UnitDivide, UnitTimes } from "./all-units";
+import * as Amount from "./amount";
+import * as Unit from "./unit";
+import * as Format from "./format";
+import * as UnitFormat from "./unit-format";
+import * as Serialize from "./serialize";
+import * as UnitPrefix from "./unit-prefix";
+import * as BaseUnits from "./base-units";
+
+export { Amount, Unit, UnitFormat, Format, Serialize, UnitPrefix, BaseUnits };
