+++ src/serialize.ts
