+++ test/amount.test.ts
@@ -1,4 +1,4 @@
-import { Amount, BaseUnits } from "../src/core";
+import { Amount, BaseUnits } from "../src";
 // import { Units } from "../src/all-units";
 import { closeTo } from "./test-utils";
